API/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
pydantic==2.4.2
uvicorn==0.23.2
google-cloud-vision==3.4.5
torch==2.1.0
transformers==4.35.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0